Remove code deprecated dust app action
