+++ front/lib/actions/utils.ts
@@ -176,7 +176,6 @@ export function getCitationsCount({
   const action = stepActions[stepActionIndex];
 
   switch (action.type) {
-    case "dust_app_run_configuration":
     case "search_labels_configuration":
       return 0;
     case "mcp_configuration":
