+++ front/lib/workspace_usage.ts
@@ -169,8 +169,6 @@ export async function unsafeGetUsageData(
     LEFT JOIN
         "agent_tables_query_configurations" atqc ON ac."id" = atqc."agentConfigurationId"
     LEFT JOIN
-        "agent_dust_app_run_configurations" adarc ON ac."id" = adarc."agentConfigurationId"
-    LEFT JOIN
         "agent_mcp_server_configurations" amsc ON ac."id" = amsc."agentConfigurationId"
     LEFT JOIN
         "mcp_server_views" msv ON amsc."mcpServerViewId" = msv."id"
